maintenance-agent/pyproject.toml
[project]
name = "maintenance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jira>=3.10.5",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "grandalf==0.8",
    "langgraph==0.6.6",
    "langgraph-checkpoint-sqlite==2.0.11",
    "langgraph-cli[inmem]==0.4.0",
    "pytest==8.4.2",
    "python-dotenv==1.1.1",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
]
